+++ release-notes.md
@@ -2,6 +2,7 @@ Version 1.8.0 (unreleased)
 --------------
 * Change the target version of the project to Java 8 to match the few
     Java 8 classes that are used.
+* Added @SafeVarargs to EntityContainer to avoid unchecked varargs warnings.
 
 
 Version 1.7.0 (latest)