+++ release-notes.md
@@ -3,8 +3,10 @@ Version 1.8.0 (unreleased)
 * Change the target version of the project to Java 8 to match the few
     Java 8 classes that are used.
 * Added @SafeVarargs to EntityContainer to avoid unchecked varargs warnings.
-* Slightly refactored ChatHostedService to make it easier for server consoles
-    to listen to and send chat messages.
+* Added ChatHostedService.add/removeChatSessionListener() so that the server
+    can listen for chat messages.
+* Added ChatHostedService.postMessage() so that a server console can post
+    messages to the chat.
 
 
 Version 1.7.0 (latest)